Cleaned up and removed some LottieComposition parsing methods
